Added generateStaticParams and shifted client side code to diffrent file.
